+++ psh/io_redirect/__init__.py
@@ -1,0 +1,4 @@
+"""I/O redirection package."""
+from .manager import IOManager
+
+__all__ = ['IOManager']